+++ tools/experimental/trt-engine-explorer/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Dates are in YYYY-MM-DD format.
 
+## v0.1.6 (2023-April)
+- Graph rendering:
+  - Add node highlighting option.
+  - Fix bug https://github.com/NVIDIA/TensorRT/issues/2779
+
 ## v0.1.5 (2022-12-06)
 - Updated requirements.txt for Ubuntu 20.04 and 22.04